Modified the theme with colors
