--- RecurRent/GlobalStyles.js
@@ -1,38 +1,0 @@
-/* fonts */
-export const FontFamily = {
-    interSemiBold: "Inter-SemiBold",
-    interRegular: "Inter-Regular",
-    interMedium: "Inter-Medium",
-    poppinsBold: "Poppins-Bold",
-};
-/* font sizes */
-export const FontSize = {
-    size_sm: 14,
-    size_base: 16,
-    size_smi: 13,
-    size_11xl: 30,
-};
-/* Colors */
-export const Color = {
-    colorWhite: "#fff",
-    colorGray_100: "#1b1d1e",
-    colorGray_200: "rgba(27, 29, 30, 0.5)",
-    colorGray_300: "rgba(27, 29, 30, 0.7)",
-    colorGainsboro: "#d8dadc",
-    colorBlack: "#000",
-    colorWhitesmoke: "#f0f0f0",
-    colorDodgerblue: "#0092ff",
-    colorCrimson: "#e64646",
-};
-/* Paddings */
-export const Padding = {
-    p_26xl: 45,
-    p_lg: 18,
-    p_133xl: 152,
-    p_mid: 17,
-    p_base: 16,
-};
-/* border radiuses */
-export const Border = {
-    br_3xs: 10,
-};
