+++ RecurRent/assets/images/adaptive-icon.png
