+++ RecurRent/package.json
@@ -9,16 +9,22 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo-google-fonts/manrope": "^0.2.3",
+    "@expo-google-fonts/montserrat": "^0.2.3",
     "@react-native-community/checkbox": "^0.5.16",
     "@react-navigation/native": "^6.1.7",
     "@react-navigation/native-stack": "^6.9.13",
     "expo": "~49.0.11",
+    "expo-font": "~11.4.0",
     "expo-image": "~1.3.3",
     "expo-status-bar": "~1.6.0",
     "firebase": "^10.4.0",
     "react": "18.2.0",
     "react-native": "0.72.4",
-    "react-native-check-box": "^2.1.7"
+    "react-native-check-box": "^2.1.7",
+    "react-native-paper": "^5.10.6",
+    "react-native-safe-area-context": "^4.7.2",
+    "react-native-vector-icons": "^10.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
