+++ RecurRent/src/screens/OpenningScreen.js
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Text, StyleSheet, View, TouchableOpacity } from "react-native";
 import { Image } from "expo-image";
-import { Padding, Border, FontFamily, FontSize, Color } from "../GlobalStyles";
+import { primaryColor, secondaryColor, textColor, backgroundColor, typography, spacing, border, lightTheme, darkTheme } from "../styles/GlobalStyles";
 
 const OpenningScreen = ({ navigation, route }) => {
   const onCreateAccountClicked = () => {
@@ -14,25 +14,25 @@ const OpenningScreen = ({ navigation, route }) => {
   return (
     <View style={styles.openningScreen}>
       <View style={styles.text}>
-        <Text style={styles.recurrent}>RecurRent</Text>
-        <Text style={styles.nowShareYour1}>
+        <Text style={[typography.heading, styles.recurrent]}>RecurRent</Text>
+        <Text style={[typography.subheading, styles.nowShareYour1]}>
           Now rent out your underutilized items and earn extra income!
         </Text>
       </View>
       <TouchableOpacity
         style={[styles.buttonPrimary1, styles.buttonFlexBox]}
         onPress={onSignInClicked}>
-        <Text style={[styles.button, styles.buttonTypo]} >Sign In</Text>
+        <Text style={[typography.body, styles.button, styles.buttonTypo]} >Sign In</Text>
       </TouchableOpacity>
       <TouchableOpacity
         style={[styles.buttonSecondary1, styles.buttonFlexBox]}
         onPress={onCreateAccountClicked}>
-        <Text style={[styles.button1, styles.buttonTypo]}>Create account</Text>
+        <Text style={[typography.body, styles.button1, styles.buttonTypo]}>Create account</Text>
       </TouchableOpacity>
       <Image
         style={styles.component4Icon}
         contentFit="cover"
-        source={require("../assets/component-4.png")}
+        source={require("../../assets/images/component-4.png")}
       />
     </View>
   );
@@ -42,44 +42,35 @@ const styles = StyleSheet.create({
   buttonFlexBox: {
     justifyContent: "center",
     alignItems: "center",
-    paddingVertical: Padding.p_mid,
+    paddingVertical: spacing.p_mid,
     // paddingHorizontal: Padding.p_133xl,
     flexDirection: "row",
     height: 56,
     width: '90%',
     marginHorizontal: '5%',
-    borderRadius: Border.br_3xs,
+    borderRadius: border.br_3xs,
     // left: 14,
     position: "absolute",
   },
   buttonTypo: {
-    fontFamily: FontFamily.interSemiBold,
     lineHeight: 20,
-    fontSize: FontSize.size_base,
     textAlign: "center",
-    fontWeight: "600",
-
   },
   recurrent: {
     top: 0,
     // left: 110,
-    fontSize: 26,
     letterSpacing: 0.8,
     lineHeight: 33,
-    fontFamily: FontFamily.poppinsSemiBold,
-    color: Color.colorBlack,
+    color: primaryColor,
     // textAlign: "left",
     textAlign: "center",
-    fontWeight: "600",
     position: "relative",
   },
   nowShareYour1: {
     top: 20,
     // left: 20,
-    fontSize: 17,
     lineHeight: 21,
-    fontFamily: FontFamily.interRegular,
-    color: "rgba(0, 53, 76, 0.7)",
+    color: secondaryColor,
     textAlign: "center",
     // width: 319,
     paddingHorizontal: '20%',
@@ -95,19 +86,19 @@ const styles = StyleSheet.create({
     // backgroundColor: "#15f5f5",
   },
   button: {
-    color: Color.colorWhite,
+    color: lightTheme.colors.onPrimaryContainer,
   },
   buttonPrimary1: {
     top: 642,
-    backgroundColor: "#0092ff",
+    backgroundColor: primaryColor,
   },
   button1: {
-    color: "#1b1d1e",
+    color: secondaryColor,
   },
   buttonSecondary1: {
     top: 712,
     borderStyle: "solid",
-    borderColor: "#b2b2b2",
+    borderColor: lightTheme.colors.outline,
     borderWidth: 1,
   },
   component4Icon: {
@@ -119,7 +110,7 @@ const styles = StyleSheet.create({
     overflow: "visible",
   },
   openningScreen: {
-    backgroundColor: Color.colorWhite,
+    backgroundColor: backgroundColor,
     flex: 1,
     width: "100%",
     // height: 852,
